Manual control code moved to joystick activity to fragment ui.
