+++ app/app/src/main/java/com/example/myapplication/EegActivity.java
@@ -48,13 +48,6 @@ public class EegActivity extends AppCompatActivity {
         // Buttons to control the start and stop of eeg reading in UI, found in content_controls.xml
         final Button controlEeg = findViewById(R.id.controlEegBtn);
 
-        // Smart car control buttons in content_controls.xml
-        final ImageButton forward = findViewById(R.id.forwardBtn);
-        // TODO: Integrate backward button into controls logic below.
-        final ImageButton backward = findViewById(R.id.backwardBtn);
-        final ImageButton left = findViewById(R.id.leftBtn);
-        final ImageButton right = findViewById(R.id.rightBtn);
-
         // Click listeners for connecting to external hardware
         btn_connectcar.setOnClickListener(new View.OnClickListener() {
             @Override
@@ -112,51 +105,6 @@ public class EegActivity extends AppCompatActivity {
                 }
             });
         }
-
-        // Click listeners for the smart car navigation control buttons
-        forward.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                try {
-                    goForward();
-                } catch (IOException e) {
-                    e.printStackTrace();
-                }
-            }
-        });
-
-        left.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                try {
-                    goLeft();
-                } catch (IOException e) {
-                    e.printStackTrace();
-                }
-            }
-        });
-
-        right.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                try {
-                    goRight();
-                } catch (IOException e) {
-                    e.printStackTrace();
-                }
-            }
-        });
-
-        backward.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                try {
-                    goBackward();
-                } catch (IOException e) {
-                    e.printStackTrace();
-                }
-            }
-        });
     }
 
     //Starts reading eeg data
@@ -256,26 +204,6 @@ public class EegActivity extends AppCompatActivity {
             tgStreamReader.startLog();
         }
         return tgStreamReader;
-    }
-
-    void goForward() throws IOException { //Buttons to steer the car
-        String msg = "c";
-        Car.mmOutputStream.write(msg.getBytes());
-    }
-
-    void goLeft() throws IOException {
-        String msg = "l";
-        Car.mmOutputStream.write(msg.getBytes());
-    }
-
-    void goRight() throws IOException {
-        String msg = "r";
-        Car.mmOutputStream.write(msg.getBytes());
-    }
-
-    void goBackward() throws IOException {
-        String msg = "b";
-        Car.mmOutputStream.write(msg.getBytes());
     }
 }
