+++ app/app/src/main/res/layout/content_controls.xml
@@ -3,18 +3,6 @@
     android:layout_width="fill_parent"
     android:layout_height="fill_parent">
 
-    <Button
-        android:id="@+id/controlEegBtn"
-        android:layout_width="100dp"
-        android:layout_height="100dp"
-        android:layout_centerHorizontal="true"
-        android:layout_centerVertical="true"
-        android:background="@drawable/bg_eegcontrol_start"
-        android:textSize="18dp"
-        android:textColor="#fff"
-        android:text="@string/start"
-        android:fontFamily="@font/nunito_extrabold" />
-
     <!-- This button is a white space filler which all other buttons align around -->
     <Button
         android:id="@+id/whitespace"
