+++ app/app/src/main/res/layout/content_header.xml
@@ -4,7 +4,7 @@
     android:id="@+id/headerView"
     android:orientation="vertical"
     android:layout_width="match_parent"
-    android:layout_height="match_parent">
+    android:layout_height="wrap_content">
 
     <TextView
         android:id="@+id/appName"
